Update cuda_kernels.cu to double

+++ cuda_kernels.cu
@@ -7,25 +7,25 @@
 #define MAX_THREADS_PER_BLOCK 1024
 
 // 数学常量
-__constant__ float c1 = -2.5f;
-__constant__ float c2 = 4.0f / 3.0f;
-__constant__ float c3 = -1.0f / 12.0f;
+__constant__ double c1 = -2.5f;
+__constant__ double c2 = 4.0f / 3.0f;
+__constant__ double c3 = -1.0f / 12.0f;
 
 // 数值稳定性常量
-__constant__ float MAX_VALUE = 1e8f;  // 最大允许值
-__constant__ float MIN_VALUE = 1e-8f; // 最小有意义值
+__constant__ double MAX_VALUE = 1e8f;  // 最大允许值
+__constant__ double MIN_VALUE = 1e-8f; // 最小有意义值
 
 // 预计算 damp1d 数组的 kernel
-__global__ void compute_damp1d_kernel(float* damp1d, float kappa, float dx, float a, int nbc) {
+__global__ void compute_damp1d_kernel(double* damp1d, double kappa, double dx, double a, int nbc) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= nbc) return;
     
-    float r = i * dx / a;
+    double r = i * dx / a;
     damp1d[i] = kappa * r * r;
 }
 
 // 吸收边界系数计算 kernel - 完全按照 Python AbcCoef2D 实现
-__global__ void abc_coef_kernel(float* abc, const float* damp1d, int nz_pad, int nx_pad, 
+__global__ void abc_coef_kernel(double* abc, const double* damp1d, int nz_pad, int nx_pad, 
                                int nbc) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     int idy = blockIdx.y * blockDim.y + threadIdx.y;
@@ -65,7 +65,7 @@ __global__ void abc_coef_kernel(float* abc, const float* damp1d, int nz_pad, int
 }
 
 // 拉普拉斯算子计算 kernel - 模拟 torch.roll 行为
-__global__ void laplacian_kernel(float* lap, const float* p, int nz_pad, int nx_pad) {
+__global__ void laplacian_kernel(double* lap, const double* p, int nz_pad, int nx_pad) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     int idy = blockIdx.y * blockDim.y + threadIdx.y;
     
@@ -90,15 +90,15 @@ __global__ void laplacian_kernel(float* lap, const float* p, int nz_pad, int nx_
     int x_minus2 = (idx - 2 + nz_pad) % nz_pad;    // torch.roll(p1, -2, 0)
     
     // 预取数据以减少内存延迟并进行数值检查
-    float p_y_plus1 = p[idx * nx_pad + y_plus1];
-    float p_y_minus1 = p[idx * nx_pad + y_minus1];
-    float p_x_plus1 = p[x_plus1 * nx_pad + idy];
-    float p_x_minus1 = p[x_minus1 * nx_pad + idy];
+    double p_y_plus1 = p[idx * nx_pad + y_plus1];
+    double p_y_minus1 = p[idx * nx_pad + y_minus1];
+    double p_x_plus1 = p[x_plus1 * nx_pad + idy];
+    double p_x_minus1 = p[x_minus1 * nx_pad + idy];
     
-    float p_y_plus2 = p[idx * nx_pad + y_plus2];
-    float p_y_minus2 = p[idx * nx_pad + y_minus2];
-    float p_x_plus2 = p[x_plus2 * nx_pad + idy];
-    float p_x_minus2 = p[x_minus2 * nx_pad + idy];
+    double p_y_plus2 = p[idx * nx_pad + y_plus2];
+    double p_y_minus2 = p[idx * nx_pad + y_minus2];
+    double p_x_plus2 = p[x_plus2 * nx_pad + idy];
+    double p_x_minus2 = p[x_minus2 * nx_pad + idy];
     
     // 数值稳定性检查
     if (!isfinite(p_y_plus1) || !isfinite(p_y_minus1) || !isfinite(p_x_plus1) ||
@@ -109,10 +109,10 @@ __global__ void laplacian_kernel(float* lap, const float* p, int nz_pad, int nx_
     }
     
     // 4阶有限差分拉普拉斯算子 - 分段计算提高精度
-    float term1 = c2 * (p_y_plus1 + p_y_minus1 + p_x_plus1 + p_x_minus1);
-    float term2 = c3 * (p_y_plus2 + p_y_minus2 + p_x_plus2 + p_x_minus2);
+    double term1 = c2 * (p_y_plus1 + p_y_minus1 + p_x_plus1 + p_x_minus1);
+    double term2 = c3 * (p_y_plus2 + p_y_minus2 + p_x_plus2 + p_x_minus2);
     
-    float lap_val = term1 + term2;
+    double lap_val = term1 + term2;
     
     // 数值范围保护
     if (isfinite(lap_val) && fabsf(lap_val) <= MAX_VALUE) {
@@ -123,9 +123,9 @@ __global__ void laplacian_kernel(float* lap, const float* p, int nz_pad, int nx_
 }
 
 // 波场更新 kernel
-__global__ void wavefield_update_kernel(float* p_new, float* p_cur, float* p_old,
-                                       const float* lap, const float* alpha, 
-                                       const float* kappa, int nz_pad, int nx_pad) {
+__global__ void wavefield_update_kernel(double* p_new, double* p_cur, double* p_old,
+                                       const double* lap, const double* alpha, 
+                                       const double* kappa, int nz_pad, int nx_pad) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     int idy = blockIdx.y * blockDim.y + threadIdx.y;
     
@@ -133,11 +133,11 @@ __global__ void wavefield_update_kernel(float* p_new, float* p_cur, float* p_old
     
     int center = idx * nx_pad + idy;
     
-    float alpha_val = alpha[center];
-    float kappa_val = kappa[center];
-    float lap_val = lap[center];
-    float p_cur_val = p_cur[center];
-    float p_old_val = p_old[center];
+    double alpha_val = alpha[center];
+    double kappa_val = kappa[center];
+    double lap_val = lap[center];
+    double p_cur_val = p_cur[center];
+    double p_old_val = p_old[center];
     
     // 启用输入值有效性检查
     if (!isfinite(alpha_val) || !isfinite(kappa_val) || !isfinite(lap_val) ||
@@ -154,15 +154,15 @@ __global__ void wavefield_update_kernel(float* p_new, float* p_cur, float* p_old
         return;
     }
     
-    float t1 = 2.0f + 2.0f * c1 * alpha_val - kappa_val;
-    float t2 = 1.0f - kappa_val;
+    double t1 = 2.0f + 2.0f * c1 * alpha_val - kappa_val;
+    double t2 = 1.0f - kappa_val;
     
     // 分段计算减少数值误差
-    float term1 = t1 * p_cur_val;
-    float term2 = t2 * p_old_val;
-    float term3 = alpha_val * lap_val;
+    double term1 = t1 * p_cur_val;
+    double term2 = t2 * p_old_val;
+    double term3 = alpha_val * lap_val;
     
-    float result = term1 - term2 + term3;
+    double result = term1 - term2 + term3;
     
     // 确保结果有效且在合理范围内
     if (isfinite(result) && fabsf(result) <= MAX_VALUE) {
@@ -174,7 +174,7 @@ __global__ void wavefield_update_kernel(float* p_new, float* p_cur, float* p_old
 }
 
 // 改进的震源注入 kernel - 增加数值保护
-__global__ void source_inject_kernel(float* p, int sz, int sx, float source_val, 
+__global__ void source_inject_kernel(double* p, int sz, int sx, double source_val, 
                                     int nx_pad) {
     // 数值范围检查
     if (!isfinite(source_val) || fabsf(source_val) > MAX_VALUE) {
@@ -184,11 +184,11 @@ __global__ void source_inject_kernel(float* p, int sz, int sx, float source_val,
     int pos = sz * nx_pad + sx;
     
     // 检查当前波场值
-    float current_val = p[pos];
+    double current_val = p[pos];
     if (!isfinite(current_val)) {
         p[pos] = source_val; // 如果当前值无效，直接设置
     } else {
-        float new_val = current_val + source_val;
+        double new_val = current_val + source_val;
         // 检查结果是否有效
         if (isfinite(new_val) && fabsf(new_val) <= MAX_VALUE) {
             p[pos] = new_val;
@@ -198,7 +198,7 @@ __global__ void source_inject_kernel(float* p, int sz, int sx, float source_val,
 }
 
 // 改进的自由表面边界条件 kernel
-__global__ void free_surface_kernel(float* p, int nbc, int nx_pad) {
+__global__ void free_surface_kernel(double* p, int nbc, int nx_pad) {
     int idy = blockIdx.x * blockDim.x + threadIdx.x;
     if (idy >= nx_pad) return;
     
@@ -210,7 +210,7 @@ __global__ void free_surface_kernel(float* p, int nbc, int nx_pad) {
         int pos_above1 = (nbc + 1) * nx_pad + idy;
         int pos_below1 = (nbc - 1) * nx_pad + idy;
         
-        float val_above1 = p[pos_above1];
+        double val_above1 = p[pos_above1];
         if (isfinite(val_above1) && fabsf(val_above1) <= MAX_VALUE) {
             p[pos_below1] = -val_above1;
         } else {
@@ -222,7 +222,7 @@ __global__ void free_surface_kernel(float* p, int nbc, int nx_pad) {
         int pos_above2 = (nbc + 2) * nx_pad + idy;
         int pos_below2 = (nbc - 2) * nx_pad + idy;
         
-        float val_above2 = p[pos_above2];
+        double val_above2 = p[pos_above2];
         if (isfinite(val_above2) && fabsf(val_above2) <= MAX_VALUE) {
             p[pos_below2] = -val_above2;
         } else {
@@ -232,7 +232,7 @@ __global__ void free_surface_kernel(float* p, int nbc, int nx_pad) {
 }
 
 // 记录提取 kernel
-__global__ void record_extract_kernel(float* seis, const float* p, 
+__global__ void record_extract_kernel(double* seis, const double* p, 
                                      const int* igz, const int* igx, 
                                      int n_rec, int nx_pad, int time_idx) {
     int rec_idx = blockIdx.x * blockDim.x + threadIdx.x;
@@ -250,23 +250,23 @@ void cuda_acoustic_forward_precomputed(torch::Tensor vpad,           // (H_pad, 
                                       torch::Tensor igx,            // (70,) 检波点 x 坐标
                                       torch::Tensor igz,            // (70,) 检波点 z 坐标
                                       torch::Tensor output,         // (5, nt, 70)
-                                      float dx, float dt, int nt, int nbc,
+                                      double dx, double dt, int nt, int nbc,
                                       bool isFS) {
     
     // 获取设备和数据类型选项
-    auto options = torch::TensorOptions().dtype(torch::kFloat32).device(vpad.device());
+    auto options = torch::TensorOptions().dtype(torch::kFloat64).device(vpad.device());
     
     int H_pad = vpad.size(0);  
     int W_pad = vpad.size(1);  
     
     // 计算吸收边界系数 - 按照 Python AbcCoef2D 实现
-    float velmin = vpad.min().item<float>();
+    double velmin = vpad.min().item<double>();
     auto abc = torch::zeros({H_pad, W_pad}, options);
     
     if (nbc > 1) {
         // 计算 kappa 和 damp1d
-        float a = (nbc - 1) * dx;
-        float kappa = 3.0f * velmin * logf(1e7f) / (2.0f * a);
+        double a = (nbc - 1) * dx;
+        double kappa = 3.0f * velmin * logf(1e7f) / (2.0f * a);
         
         // 分配 damp1d 数组
         auto damp1d = torch::zeros({nbc}, options);
@@ -275,7 +275,7 @@ void cuda_acoustic_forward_precomputed(torch::Tensor vpad,           // (H_pad, 
         dim3 block_1d(256);
         dim3 grid_1d((nbc + block_1d.x - 1) / block_1d.x);
         compute_damp1d_kernel<<<grid_1d, block_1d>>>(
-            damp1d.data_ptr<float>(), kappa, dx, a, nbc);
+            damp1d.data_ptr<double>(), kappa, dx, a, nbc);
         
         // 计算 ABC 系数
         dim3 block_2d(BLOCK_SIZE, BLOCK_SIZE);
@@ -283,7 +283,7 @@ void cuda_acoustic_forward_precomputed(torch::Tensor vpad,           // (H_pad, 
                      (W_pad + block_2d.y - 1) / block_2d.y);
         
         abc_coef_kernel<<<grid_2d, block_2d>>>(
-            abc.data_ptr<float>(), damp1d.data_ptr<float>(),
+            abc.data_ptr<double>(), damp1d.data_ptr<double>(),
             H_pad, W_pad, nbc);
     }
     
@@ -320,33 +320,33 @@ void cuda_acoustic_forward_precomputed(torch::Tensor vpad,           // (H_pad, 
         for (int it = 0; it < nt; it++) {
             // 计算拉普拉斯算子
             laplacian_kernel<<<grid_2d, block_2d>>>(
-                lap.data_ptr<float>(), p_cur.data_ptr<float>(),
+                lap.data_ptr<double>(), p_cur.data_ptr<double>(),
                 H_pad, W_pad);
             
             // 更新波场
             wavefield_update_kernel<<<grid_2d, block_2d>>>(
-                p_new.data_ptr<float>(), p_cur.data_ptr<float>(),
-                p_old.data_ptr<float>(), lap.data_ptr<float>(),
-                alpha.data_ptr<float>(), kappa.data_ptr<float>(),
+                p_new.data_ptr<double>(), p_cur.data_ptr<double>(),
+                p_old.data_ptr<double>(), lap.data_ptr<double>(),
+                alpha.data_ptr<double>(), kappa.data_ptr<double>(),
                 H_pad, W_pad);
             
             // 注入震源
-            float source_val = beta_d[isz_val][sx].item<float>() * wavelet[it].item<float>();
+            double source_val = beta_d[isz_val][sx].item<double>() * wavelet[it].item<double>();
             source_inject_kernel<<<1, 1>>>(
-                p_new.data_ptr<float>(), isz_val, sx, source_val, W_pad);
+                p_new.data_ptr<double>(), isz_val, sx, source_val, W_pad);
             
             // 自由表面边界条件
             if (isFS) {
                 dim3 grid_fs((W_pad + 255) / 256);
                 free_surface_kernel<<<grid_fs, 256>>>(
-                    p_new.data_ptr<float>(), nbc, W_pad);
+                    p_new.data_ptr<double>(), nbc, W_pad);
             }
             
             // 记录地震数据
             dim3 grid_rec((n_rec + 255) / 256);
             auto seis_slice = output.select(0, src_idx);  // (nt, 70)
             record_extract_kernel<<<grid_rec, 256>>>(
-                seis_slice.select(0, it).data_ptr<float>(), p_new.data_ptr<float>(),
+                seis_slice.select(0, it).data_ptr<double>(), p_new.data_ptr<double>(),
                 igz.data_ptr<int>(), igx.data_ptr<int>(),
                 n_rec, W_pad, 0);
 
@@ -370,7 +370,7 @@ void cuda_acoustic_forward_batch(torch::Tensor vpad_batch,       // (B, H_pad, W
                                  torch::Tensor igx,              // (70,) 检波点 x 坐标
                                  torch::Tensor igz,              // (70,) 检波点 z 坐标
                                  torch::Tensor output_batch,     // (B, 5, nt, 70)
-                                 float dx, float dt, int nt, int nbc,
+                                 double dx, double dt, int nt, int nbc,
                                  bool isFS) {
     
     // 获取批次大小和空间维度
@@ -382,7 +382,7 @@ void cuda_acoustic_forward_batch(torch::Tensor vpad_batch,       // (B, H_pad, W
     int isz_val = isz.item<int>();
     
     // 获取设备和数据类型选项
-    auto options = torch::TensorOptions().dtype(torch::kFloat32).device(vpad_batch.device());
+    auto options = torch::TensorOptions().dtype(torch::kFloat64).device(vpad_batch.device());
     
     // 创建CUDA streams for 并行处理
     const int MAX_STREAMS = 8;  // 最大并发流数量
@@ -443,13 +443,13 @@ void cuda_acoustic_forward_batch(torch::Tensor vpad_batch,       // (B, H_pad, W
         beta_d.mul_(dt).pow_(2);
         
         // 计算吸收边界系数
-        float velmin = vpad_current.min().item<float>();
+        double velmin = vpad_current.min().item<double>();
         abc.zero_();
         
         if (nbc > 1) {
             // 计算 kappa 和 damp1d
-            float a = (nbc - 1) * dx;
-            float kappa_val = 3.0f * velmin * logf(1e7f) / (2.0f * a);
+            double a = (nbc - 1) * dx;
+            double kappa_val = 3.0f * velmin * logf(1e7f) / (2.0f * a);
             
             // 分配 damp1d 数组（每个stream独立）
             auto damp1d = torch::zeros({nbc}, options);
@@ -457,11 +457,11 @@ void cuda_acoustic_forward_batch(torch::Tensor vpad_batch,       // (B, H_pad, W
             // 在指定stream中异步执行
             dim3 grid_1d((nbc + block_1d.x - 1) / block_1d.x);
             compute_damp1d_kernel<<<grid_1d, block_1d, 0, current_stream>>>(
-                damp1d.data_ptr<float>(), kappa_val, dx, a, nbc);
+                damp1d.data_ptr<double>(), kappa_val, dx, a, nbc);
             
             // 计算 ABC 系数
             abc_coef_kernel<<<grid_2d, block_2d, 0, current_stream>>>(
-                abc.data_ptr<float>(), damp1d.data_ptr<float>(),
+                abc.data_ptr<double>(), damp1d.data_ptr<double>(),
                 H_pad, W_pad, nbc);
         }
         
@@ -484,31 +484,31 @@ void cuda_acoustic_forward_batch(torch::Tensor vpad_batch,       // (B, H_pad, W
                 
                 // 计算拉普拉斯算子
                 laplacian_kernel<<<grid_2d, block_2d, 0, current_stream>>>(
-                    lap.data_ptr<float>(), p_cur.data_ptr<float>(),
+                    lap.data_ptr<double>(), p_cur.data_ptr<double>(),
                     H_pad, W_pad);
                 
                 // 更新波场
                 wavefield_update_kernel<<<grid_2d, block_2d, 0, current_stream>>>(
-                    p_new.data_ptr<float>(), p_cur.data_ptr<float>(),
-                    p_old.data_ptr<float>(), lap.data_ptr<float>(),
-                    alpha.data_ptr<float>(), kappa.data_ptr<float>(),
+                    p_new.data_ptr<double>(), p_cur.data_ptr<double>(),
+                    p_old.data_ptr<double>(), lap.data_ptr<double>(),
+                    alpha.data_ptr<double>(), kappa.data_ptr<double>(),
                     H_pad, W_pad);
                 
                 // 注入震源
-                float source_val = beta_d[isz_val][sx].item<float>() * wavelet[it].item<float>();
+                double source_val = beta_d[isz_val][sx].item<double>() * wavelet[it].item<double>();
                 source_inject_kernel<<<1, 1, 0, current_stream>>>(
-                    p_new.data_ptr<float>(), isz_val, sx, source_val, W_pad);
+                    p_new.data_ptr<double>(), isz_val, sx, source_val, W_pad);
                 
                 // 自由表面边界条件
                 if (isFS) {
                     free_surface_kernel<<<grid_fs, 256, 0, current_stream>>>(
-                        p_new.data_ptr<float>(), nbc, W_pad);
+                        p_new.data_ptr<double>(), nbc, W_pad);
                 }
                 
                 // 记录地震数据
                 auto seis_slice = output_current.select(0, src_idx);  // (nt, 70)
                 record_extract_kernel<<<grid_rec, 256, 0, current_stream>>>(
-                    seis_slice.select(0, it).data_ptr<float>(), p_new.data_ptr<float>(),
+                    seis_slice.select(0, it).data_ptr<double>(), p_new.data_ptr<double>(),
                     igz.data_ptr<int>(), igx.data_ptr<int>(),
                     n_rec, W_pad, 0);
                 
@@ -536,7 +536,7 @@ void cuda_acoustic_forward_batch_optimized(torch::Tensor vpad_batch,       // (B
                                            torch::Tensor igx,              // (70,) 检波点 x 坐标
                                            torch::Tensor igz,              // (70,) 检波点 z 坐标
                                            torch::Tensor output_batch,     // (B, 5, nt, 70)
-                                           float dx, float dt, int nt, int nbc,
+                                           double dx, double dt, int nt, int nbc,
                                            bool isFS, int max_concurrent_batches = 4) {
     
     int B = vpad_batch.size(0);
